install, comm: remove file_diff in favor of uucore::fs::are_files_identical
