+++ src/uu/install/Cargo.toml
@@ -16,7 +16,6 @@ path = "src/install.rs"
 
 [dependencies]
 clap = { workspace = true }
-file_diff = { workspace = true }
 thiserror = { workspace = true }
 uucore = { workspace = true, default-features = true, features = [
   "backup-control",
